// src/sendbird.d.ts
export interface BaseMessageInstance {
    channelUrl: string;
    channelType: string;
    messageId: number;
    message: string;
    messageType: string;
    data: string;
    customType: string;
    createdAt: number;
    updatedAt: number;

    isOpenChannel(): boolean;
    isGroupChannel(): boolean;
    isUserMessage(): boolean;
    isFileMessage(): boolean;
    isAdminMessage(): boolean;
    serialize(): Object;
}

export interface AdminMessage extends BaseMessageInstance {
    translations: Object;
}
export interface AdminMessageStatic {
    buildFromSerializedData(serializedObject: Object): AdminMessage;
}

export interface UserMessage extends BaseMessageInstance {
    sender: User;
    reqId: string;
    translations: Object;
}
export interface UserMessageStatic {
    buildFromSerializedData(serializedObject: Object): UserMessage;
}

export interface FileMessage extends BaseMessageInstance {
    sender: User;
    reqId: string;
    url: string;
    name: string;
    size: number;
    type: string;
    thumbnails: Array<ThumbnailObject>;
}
export interface FileMessageStatic {
    buildFromSerializedData(serializedObject: Object): FileMessage;
}

export interface ThumbnailObject {
    url: string;
    height: number;
    width: number;
    real_height: number;
    real_width: number;
}
export interface ThumbnailSize {
    maxWidth: number;
    maxHeight: number;
}

/**
 * User
 */
export interface User {
    userId: string;
    nickname: string;
    profileUrl: string;
    metaData: Object;
    connectionStatus: string;
    lastSeenAt: string;

    createMetaData(metaDataMap: Object, callback: commonCallback): void;
    updateMetaData(metaDataMap: Object, callback: commonCallback): void;
    updateMetaData(metaDataMap: Object, upsert: boolean, callback: commonCallback): void;
    deleteMetaData(metaDataKey: string, callback: commonCallback): void;
    deleteAllMetaData(callback: commonCallback): void;

    serialize(): Object;
}
export interface UserStatic {
    buildFromSerializedData(serializedObject: Object): User;
}

export interface Member extends User {
    state: 'invited' | 'joined';
    isBlockedByMe: boolean;
    isBlockingMe: boolean;
}
// ...
export type commonCallback = (response: Object, error: Object) => void;
export type openChannelCallback = (openChannel: OpenChannel, error: Object) => void;
export interface OpenChannel extends BaseChannel {
    isFrozen: boolean;
    participantCount: number;
    operators: Array<User>;

    refresh(callback: openChannelCallback): void;
    delete(callback: openChannelCallback): void;
    enter(callback: openChannelCallback): void;
    exit(callback: openChannelCallback): void;

    createParticipantListQuery(): UserListQuery;
    createMutedUserListQuery(): UserListQuery;
    createBannedUserListQuery(): UserListQuery;

    updateChannel(name: string, coverUrl: string, data: string, callback: openChannelCallback): void;
    updateChannel(name: string, coverUrl: string, data: string, operatorUserIds: Array<string> | string, callback: openChannelCallback): void;
    updateChannel(name: string, coverUrl: string, data: string, operatorUserIds: Array<string> | string, customType: string, callback: openChannelCallback): void;

    updateChannelWithOperatorUserIds(name: string, coverUrl: string, data: string, operatorUserIds: Array<string> | string, callback: openChannelCallback): void;
    updateChannelWithOperatorUserIds(name: string, coverUrl: string, data: string, operatorUserIds: Array<string> | string, customType: string, callback: openChannelCallback): void;

    banUser(user: User, callback: commonCallback): void;
    banUser(user: User, seconds: number, callback: commonCallback): void;
    banUserWithUserId(userId: string, callback: commonCallback): void;
    banUserWithUserId(userId: string, seconds: number, callback: commonCallback): void;
    unbanUser(user: User, callback: commonCallback): void;
    unbanUserWithUserId(userId: string, callback: commonCallback): void;

    muteUser(user: User, callback: commonCallback): void;
    muteUserWithUserId(userId: string, callback: commonCallback): void;
    unmuteUser(user: User, callback: commonCallback): void;
    unmuteUserWithUserId(userId: string, callback: commonCallback): void;

    isOperator(user: User): boolean;
    isOperatorWithUserId(userId: string): boolean;
}

export type userListQueryCallback = (userList: Array<User>, error: Object) => void;
export interface UserListQuery {
    limit: number;
    hasNext: boolean;
    isLoading: boolean;
    metaDataKey: string;
    metaDataValues: Array<string>;

    next(callback: userListQueryCallback): void;
}
// ...
type groupChannelCallback = (groupChannel: GroupChannel, error: Object) => void;
type getPushPreferenceCallback = (isPushOn: boolean, error: Object) => void;
export interface GroupChannel extends BaseChannel {
    isDistinct: boolean;
    isPushEnabled: boolean;
    lastMessage: UserMessage | FileMessage | AdminMessage;
    unreadMessageCount: number;
    members: Array<Member>;
    memberCount: number;
    inviter: User;

    refresh(callback: groupChannelCallback): void;

    updateChannel(name: string, coverUrlOrImageFile: string | File, data: string, callback: groupChannelCallback): void;
    updateChannel(isDistinct: boolean, name: string, coverUrlOrImageFile: string | File, data: string, callback: groupChannelCallback): void;
    updateChannel(isDistinct: boolean, name: string, coverUrlOrImageFile: string | File, data: string, customType: string, callback: groupChannelCallback): void;

    resetMyHistory(callback?: commonCallback): void;

    invite(users: Array<User>, callback: groupChannelCallback): void;
    inviteWithUserIds(userIds: Array<string>, callback: groupChannelCallback): void;
    acceptInvitation(callback: groupChannelCallback): void;
    declineInvitation(callback: commonCallback): void;

    leave(callback: commonCallback): void;

    hide(callback: commonCallback): void;
    hide(hidePreviousMessages: boolean, callback: commonCallback): void;

    markAsRead(): void;
    getReadReceipt(message: UserMessage | FileMessage | AdminMessage): number;
    getReadStatus(): Object;

    startTyping(): void;
    endTyping(): void;
    isTyping(): boolean;
    getTypingMembers(): Array<Member>;

    setPushPreference(pushOn: boolean, callback: commonCallback): void;
    getPushPreference(callback: getPushPreferenceCallback): void;

}